+++ HISTORY.rst
@@ -8,7 +8,7 @@ History
 
 * Support for Django 1.6 and lower is now deprecated.
 * Improved test harness now tests coverage and pep8
-* SubscribeFormView no longer populates self.error with form errors
+* SubscribeFormView and ChangePlanView no longer populate self.error with form errors
 
 0.5.0 (2015-05-25)
 ---------------------